requirements.txt
colorama==0.4.6
numpy==2.1.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow==10.4.0
PyYAML==6.0.2
tqdm==4.66.5
